Modularise cheerio functions
