+++ models/cheerio/remove-empty-paragraphs.js
@@ -1,0 +1,8 @@
+
+/**
+ * Removes paragraphs elements containing only whitespace nodes
+ * @return String
+ */
+module.exports = function(html) {
+    return html.replace(/<p>\s+<\/p>/, '');
+};
